Develop

+++ migrations/008.cretae_table_bank_account.sql
@@ -3,7 +3,7 @@ CREATE TABLE if not exists Bank_Account(
     id uuid DEFAULT uuid_generate_v4 (),
     currencyId integer,
     userId integer,
-    value integer,
+    value float,
     FOREIGN KEY (currencyId) REFERENCES Currency(id),
     FOREIGN KEY (userId) REFERENCES Users(id)
 )
